backend/requirements.txt
flask>=2.0
flask-cors>=3.0
opencv-python-headless>=4.7
numpy>=1.22
tensorflow>=2.10
keras>=2.10
cvzone>=1.5
mediapipe>=0.9
pyttsx3>=2.90
pyngrok>=5.1.0
requests>=2.28
pyenchant>=3.2
